reorganize org.cfg4j.source package
