+++ cfg4j-core/src/main/java/org/cfg4j/source/ConfigurationSource.java
@@ -15,8 +15,8 @@
  */
 package org.cfg4j.source;
 
-import org.cfg4j.source.context.Environment;
-import org.cfg4j.source.context.MissingEnvironmentException;
+import org.cfg4j.source.context.environment.Environment;
+import org.cfg4j.source.context.environment.MissingEnvironmentException;
 import org.cfg4j.source.reload.Reloadable;
 
 import java.util.Properties;
