+++ cfg4j-core/src/test/java/org/cfg4j/source/compose/FallbackConfigurationSourceTest.java
@@ -26,8 +26,8 @@ import static org.mockito.Mockito.when;
 
 import org.assertj.core.data.MapEntry;
 import org.cfg4j.source.ConfigurationSource;
-import org.cfg4j.source.context.Environment;
-import org.cfg4j.source.context.MissingEnvironmentException;
+import org.cfg4j.source.context.environment.Environment;
+import org.cfg4j.source.context.environment.MissingEnvironmentException;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
