+++ cfg4j-core/src/test/java/org/cfg4j/source/context/environment/ImmutableEnvironmentTest.java
@@ -13,31 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cfg4j.source;
+package org.cfg4j.source.context.environment;
 
-import org.cfg4j.source.context.Environment;
+import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.Properties;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.runners.MockitoJUnitRunner;
 
-/**
- * Empty {@link ConfigurationSource}
- */
-public class EmptyConfigurationSource implements ConfigurationSource {
 
-  private static final Properties properties = new Properties();
+@RunWith(MockitoJUnitRunner.class)
+public class ImmutableEnvironmentTest {
 
-  @Override
-  public Properties getConfiguration(Environment environment) {
-    return properties;
-  }
-
-  @Override
-  public void reload() {
-    // NOP
-  }
-
-  @Override
-  public String toString() {
-    return "EmptyConfigurationSource{}";
+  @Test
+  public void shouldReturnConstantEnvName() throws Exception {
+    String environmentName = "sampleEnvironment";
+    ImmutableEnvironment strategy = new ImmutableEnvironment(environmentName);
+    assertThat(strategy.getName()).isEqualTo(environmentName);
   }
 }
