+++ cfg4j-core/src/test/java/org/cfg4j/source/context/filesprovider/DefaultConfigFilesProviderTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cfg4j.source;
+package org.cfg4j.source.context.filesprovider;
 
 import static org.assertj.core.api.Assertions.assertThat;
