+++ cfg4j-core/src/test/java/org/cfg4j/source/empty/EmptyConfigurationSourceTest.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package org.cfg4j.source;
+package org.cfg4j.source.empty;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
-import org.cfg4j.source.context.DefaultEnvironment;
-import org.cfg4j.source.context.Environment;
+import org.cfg4j.source.context.environment.DefaultEnvironment;
+import org.cfg4j.source.context.environment.Environment;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
