+++ cfg4j-core/src/test/java/org/cfg4j/source/files/FilesConfigurationSourceTest.java
@@ -18,11 +18,11 @@ package org.cfg4j.source.files;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.assertj.core.data.MapEntry;
-import org.cfg4j.source.ConfigFilesProvider;
-import org.cfg4j.source.context.DefaultEnvironment;
-import org.cfg4j.source.context.Environment;
-import org.cfg4j.source.context.ImmutableEnvironment;
-import org.cfg4j.source.context.MissingEnvironmentException;
+import org.cfg4j.source.context.environment.DefaultEnvironment;
+import org.cfg4j.source.context.environment.Environment;
+import org.cfg4j.source.context.environment.ImmutableEnvironment;
+import org.cfg4j.source.context.environment.MissingEnvironmentException;
+import org.cfg4j.source.context.filesprovider.ConfigFilesProvider;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
