+++ cfg4j-git/src/main/java/org/cfg4j/source/git/GitConfigurationSource.java
@@ -17,11 +17,11 @@ package org.cfg4j.source.git;
 
 import static java.util.Objects.requireNonNull;
 
-import org.cfg4j.source.ConfigFilesProvider;
 import org.cfg4j.source.ConfigurationSource;
 import org.cfg4j.source.SourceCommunicationException;
-import org.cfg4j.source.context.Environment;
-import org.cfg4j.source.context.MissingEnvironmentException;
+import org.cfg4j.source.context.environment.Environment;
+import org.cfg4j.source.context.environment.MissingEnvironmentException;
+import org.cfg4j.source.context.filesprovider.ConfigFilesProvider;
 import org.cfg4j.utils.FileUtils;
 import org.cfg4j.utils.PropertiesProvider;
 import org.cfg4j.utils.PropertiesProviderSelector;
